Prevent overwrite file

+++ install.sh
@@ -36,6 +36,11 @@ cp ./tmux/tmux.conf ~/.tmux.conf
 
 # Git
 
-cp ./git/gitconfig ~/.gitconfig
-cp ./git/gitignore ~/.gitignore
-git config --global core.excludesfile ~/.gitignore
+if [ ! -f ~/.gitconfig ]; then
+	cp ./git/gitconfig ~/.gitconfig
+fi 
+
+if [ ! -f ~/.gitignore ]; then
+	cp ./git/gitignore ~/.gitignore
+	git config --global core.excludesfile ~/.gitignore
+fi
